🚚 move `utils_test.py` in `./app/test` folder
